# PERF: improves SeriesGroupBy.nunique performance