@behzadnouri 2015-08-24
benchmarks:

```
-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------
groupby_ngroups_10000_nunique                |   6.8550 | 1701.2453 |   0.0040 |
groupby_ngroups_100_nunique                  |   0.5283 |  18.2413 |   0.0290 |
-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------

Ratio < 1.0 means the target commit is faster then the baseline.
Seed used: 1234

Target [091c73d] : PERF: improves SeriesGroupBy.nunique performance
Base   [54f02df] : COMPAT: value_counts always return int64 dtype, xref #10876
```