@aldanor 2015-08-24
That's awesome, sort of what I was thinking about but didn't have time to hack together :/ Thanks @behzadnouri!